LAB box plot from CSV test data, based on VS boxplot code
Created WPCT-F.07.01-sdtm-lb.sas template program as a test of:
1. producing a LAB box plot from the VS SAS code, and
2. producing a plot from the SDTM CSV data rather than the ADaM XPT
data.

This required a small change to UTIL_LABELS_FROM_VAR.SAS to avoid
overwriting macro variables that provide labels for PARAMCDs. We need
two sets for labs: 1st for Long Names, 2nd for Units.

Also added SAS programs to import the CSV data from github, since the
LRECL settings are subtle and essential for the import to work.