Fig. 7.7 implemented in SAS
Including, in addition to SAS program WPCT-F.07.07.sas:
+ docx specs, WPCT_Fig_7.7_RequirementsSpecification.docx
+ YML, WPCT-F.07.07_sas.yml
Minor changes to Fig. 7.6 specs, and SAS programs for Figs. 7.2 and 7.3,
for consistency.

LAST/MIN/MAX requires derivation, which should not be in a display
program. So I isolated this as and example PhUSE CS ADaM derivation in
program whitepapers\ADaM\derive_lastminmax_measure.sas