Create ae_meddra.sas
ae_meddra_w_flag_generation_v1.sas relies on this macro, but I cannot find it anywhere in the "tested" area. I downloaded it from the "contributed" area and suggest we add it here so that it resides with the other macro, utility, and template files for the JumpStart scripts.